vhd: convert vmdk-to-vhd tests to jest.
