+++ .travis.yml
@@ -12,6 +12,7 @@ addons:
     packages:
     - qemu-utils
     - blktap-utils
+    - vmdk-stream-converter
 
 before_install:
   - curl -o- -L https://yarnpkg.com/install.sh | bash
