+++ packages/xo-vmdk-to-vhd/.babelrc.js
@@ -1,0 +1,3 @@
+module.exports = require('../../@xen-orchestra/babel-config')(
+  require('./package.json')
+)
