+++ packages/xo-vmdk-to-vhd/package.json
@@ -30,37 +30,28 @@
     "pipette": "^0.9.3"
   },
   "devDependencies": {
-    "babel-cli": "^6.18.0",
-    "babel-plugin-transform-runtime": "^6.15.0",
-    "babel-preset-env": "^1.0.0",
+    "@babel/cli": "7.0.0-beta.42",
+    "@babel/core": "7.0.0-beta.42",
+    "@babel/plugin-proposal-decorators": "7.0.0-beta.42",
+    "@babel/plugin-proposal-export-default-from": "7.0.0-beta.42",
+    "@babel/plugin-proposal-export-namespace-from": "7.0.0-beta.42",
+    "@babel/plugin-proposal-function-bind": "7.0.0-beta.42",
+    "@babel/plugin-proposal-optional-chaining": "^7.0.0-beta.40",
+    "@babel/plugin-proposal-pipeline-operator": "^7.0.0-beta.40",
+    "@babel/plugin-proposal-throw-expressions": "^7.0.0-beta.40",
+    "@babel/preset-env": "7.0.0-beta.42",
+    "@babel/preset-flow": "7.0.0-beta.42",
+    "babel-plugin-lodash": "^3.3.2",
     "cross-env": "^5.1.3",
-    "mocha": "^4.0.1",
-    "must": "^0.13.2",
-    "rimraf": "^2.5.4"
+    "index-modules": "^0.3.0",
+    "rimraf": "^2.6.2"
   },
   "scripts": {
     "build": "cross-env NODE_ENV=production babel --source-maps --out-dir=dist/ src/",
     "clean": "rimraf dist/",
     "dev": "cross-env NODE_ENV=development babel --watch --source-maps --out-dir=dist/ src/",
-    "dev-test": "mocha --watch --reporter=min \"dist/**/*.spec.js\"",
     "prebuild": "yarn run clean",
-    "predev": "yarn run prebuild",
-    "prepublishOnly": "yarn run build",
-    "test-disabled": "mocha \"dist/**/*.spec.js\""
-  },
-  "babel": {
-    "plugins": [
-      "transform-runtime"
-    ],
-    "presets": [
-      [
-        "env",
-        {
-          "targets": {
-            "node": 4
-          }
-        }
-      ]
-    ]
+    "predev": "yarn run clean",
+    "prepublishOnly": "yarn run build"
   }
 }
